pyproject.toml
[project]
name = "mcp-tool-web-search"
version = "0.1.9"
description = "MCP server for web search - 将网络搜索能力封装为 Model Context Protocol 工具"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "XiaoHui2023"}]
license = { text = "MIT" }
keywords = ["mcp", "model-context-protocol", "web-search", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.2.0",
    "pydantic-settings>=2.0.0",
    "curl-cffi>=0.7.0",
    "trafilatura>=1.6.0",
    "requests>=2.31.0",
    "python-frontmatter>=1.0.0",   # 新增
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.7",
]

[project.scripts]
mcp-tool-web-search = "mcp_tool_web_search.__main__:main"

[project.urls]
Homepage = "https://github.com/XiaoHui2023/mcp-tool-web-search"
Repository = "https://github.com/XiaoHui2023/mcp-tool-web-search"
Documentation = "https://github.com/XiaoHui2023/mcp-tool-web-search#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_tool_web_search"]